requirements.txt
torch==1.9.1
pandas==1.3.4
opencv_python==4.5.3.56
numpy==1.21.2
joblib==1.0.1
torchvision==0.10.1
Pillow==8.3.1
scikit_learn==0.24.2
